Module etl_job::job

Create pipelines linking etl_core::datastore::DataSources together; Provides state management based on a job id, responsible for generating and running pipelines, including any run-time configurations

Modules

command

error

handler

state

stream

stream_handler_builder

Structs

JobItemInfo

JobRunner

JobRunner is used to declare pipelines which are designed to run step-by-step. To execute multiple pipelines in parallel, use separate JobRunner for each pipeline.

JobRunnerConfig

Enums

JobRunnerAction